feat: add security to endpoints (#18)
* hotfix: update gitignore

* hotfix: update docker-compose file - remove env-files config

* hotfix: structured and enhancement project (#3)

* hotfix: update readme project

* feat: add new environment variable

* feat: add package beanie-odm

* fix: remove unsing config in docker-compose file

* feat: update botoclient function

* feat: cleanup error codes and add news

* feat: remove unusing functions and update utils functionnal

* feat: define bucket policy config

* test: add tests for endpoints and functional (#5)

* test: restructured test config, add test to media endpoint (#7)

* test: add tests for endpoints and functional

* test: restructured test config, add test to media endpoint

* feat: add workflow config (#9)

* feat: update workflow (#10)

* feat: add workflow config

* feat: add github action

* fix: fix workflow ci (#11)

* Fix/worklows (#12)

* fix: fix workflow ci

* fix: fix workflow ci

* Fix/worklows (#13)

* fix: fix workflow ci

* fix: fix workflow ci

* fix: create docker image (#15)

* fix: fix workflow ci

* fix: fix workflow ci

* fix: environment

* feat: add permissions to endpoint

+++ tests/routers/test_bucket_api.py
@@ -16,7 +16,7 @@ async def test_ping_api(http_client_api):
 
 
 @pytest.mark.asyncio
-async def test_create_bucket_success(http_client_api, mock_boto_client, default_bucket, bucket_data):
+async def test_create_bucket_success(http_client_api, mock_boto_client, default_bucket, mock_check_access_allow, bucket_data):
     bucket_data.update({"bucket_name": "storage-unsta-pictures"})
 
     # Simuler que le bucket n'existe pas en levant une exception 404
@@ -28,7 +28,7 @@ async def test_create_bucket_success(http_client_api, mock_boto_client, default_
     mock_boto_client.create_bucket.return_value = {}
     mock_boto_client.put_bucket_policy.return_value = {}
 
-    response = await http_client_api.post("/buckets", json=bucket_data)
+    response = await http_client_api.post("/buckets", json=bucket_data, headers={"Authorization": "Bearer token"})
 
     assert response.status_code == status.HTTP_201_CREATED, response.text
     assert response.json()["bucket_name"] == bucket_data.get("bucket_name")
@@ -40,14 +40,16 @@ async def test_create_bucket_success(http_client_api, mock_boto_client, default_
         Bucket=bucket_data.get("bucket_name"), CreateBucketConfiguration={"LocationConstraint": settings.STORAGE_REGION_NAME}
     )
 
+    mock_check_access_allow.assert_called_once()
 
-async def test_create_bucket_failure(http_client_api, mock_boto_client, bucket_data):
+
+async def test_create_bucket_failure(http_client_api, mock_boto_client, bucket_data, mock_check_access_allow):
     mock_boto_client.head_bucket.side_effect = exceptions.ClientError(
         error_response={"Error": {"Code": "400", "Message": "Error checking bucket"}}, operation_name="HeadBucket"
     )
 
     bucket_data.update({"bucket_name": "storage-pictures"})
-    response = await http_client_api.post("/buckets", json=bucket_data)
+    response = await http_client_api.post("/buckets", json=bucket_data, headers={"Authorization": "Bearer token"})
 
     assert response.status_code == status.HTTP_400_BAD_REQUEST, response.text
     assert response.json()["error_code"] == SfsErrorCodes.SFS_INVALID_NAME
@@ -56,44 +58,55 @@ async def test_create_bucket_failure(http_client_api, mock_boto_client, bucket_d
     )
 
     mock_boto_client.head_bucket.assert_called_once_with(Bucket="storage-pictures")
+    mock_check_access_allow.assert_called_once()
 
 
 @pytest.mark.asyncio
-async def test_list_bucket_with_data(http_client_api, default_bucket):
-    response = await http_client_api.get("/buckets")
+async def test_list_bucket_with_data(http_client_api, default_bucket, mock_check_access_allow):
+    response = await http_client_api.get("/buckets", headers={"Authorization": "Bearer token"})
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json()["total"] >= 1
     assert response.json()["items"][0]["bucket_name"] == default_bucket.bucket_name
 
+    mock_check_access_allow.assert_called_once()
+
 
 @pytest.mark.asyncio
-async def test_list_bucket_without_data(http_client_api):
-    response = await http_client_api.get("/buckets")
+async def test_list_bucket_without_data(http_client_api, mock_check_access_allow):
+    response = await http_client_api.get("/buckets", headers={"Authorization": "Bearer token"})
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json() == {"items": [], "total": 0, "page": 1, "size": None, "pages": 0}
 
+    mock_check_access_allow.assert_called_once()
+
 
 @pytest.mark.asyncio
-async def test_list_buckect_filter(http_client_api, default_bucket, fake_data):
+async def test_list_buckect_filter(http_client_api, default_bucket, fake_data, mock_check_access_allow):
     # Test filter by bucket_name
-    response = await http_client_api.get("/buckets", params={"bucket_name": f"{default_bucket.bucket_name}"})
+    response = await http_client_api.get(
+        "/buckets", params={"bucket_name": f"{default_bucket.bucket_name}"}, headers={"Authorization": "Bearer token"}
+    )
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json()["items"][0]["bucket_name"] == default_bucket.bucket_name
 
     # Test filter by description
-    response = await http_client_api.get("/buckets", params={"description": default_bucket.description})
+    response = await http_client_api.get(
+        "/buckets", params={"description": default_bucket.description}, headers={"Authorization": "Bearer token"}
+    )
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json()["items"][0]["description"] == default_bucket.description
 
     # Test filter by created_at
     query_date = fake_data.date()
-    response = await http_client_api.get("/buckets", params={"created_at": query_date})
+    response = await http_client_api.get("/buckets", params={"created_at": query_date}, headers={"Authorization": "Bearer token"})
     assert response.status_code == status.HTTP_200_OK, response.text
     assert "items" in response.json()
 
+    mock_check_access_allow.assert_called()
+
 
 @pytest.mark.asyncio
-async def test_get_bucket_and_create_bucket_if_not_exist(http_client_api, mock_boto_client):
+async def test_get_bucket_and_create_bucket_if_not_exist(http_client_api, mock_boto_client, mock_check_access_allow):
     # Simuler que le bucket n'existe pas en levant une exception 404
     mock_boto_client.head_bucket.side_effect = exceptions.ClientError(
         error_response={"Error": {"Code": "404", "Message": "Not Found"}}, operation_name="HeadBucket"
@@ -103,7 +116,9 @@ async def test_get_bucket_and_create_bucket_if_not_exist(http_client_api, mock_b
     mock_boto_client.create_bucket.return_value = {}
     mock_boto_client.put_bucket_policy.return_value = {}
 
-    response = await http_client_api.get("/buckets/unsta-pictures", params={"create_bucket_if_not_exist": True})
+    response = await http_client_api.get(
+        "/buckets/unsta-pictures", params={"create_bucket_if_not_exist": True}, headers={"Authorization": "Bearer token"}
+    )
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json()["bucket_name"] == "unsta-pictures"
 
@@ -114,22 +129,32 @@ async def test_get_bucket_and_create_bucket_if_not_exist(http_client_api, mock_b
         Bucket="unsta-pictures", CreateBucketConfiguration={"LocationConstraint": settings.STORAGE_REGION_NAME}
     )
 
+    mock_check_access_allow.assert_called_once()
+
 
 @pytest.mark.asyncio
-async def test_get_bucket_and_create_bucket_if_exist(http_client_api, mock_boto_client, default_bucket):
-    response = await http_client_api.get(f"/buckets/{default_bucket.bucket_name}", params={"create_bucket_if_not_exist": False})
+async def test_get_bucket_and_create_bucket_if_exist(http_client_api, mock_boto_client, default_bucket, mock_check_access_allow):
+    response = await http_client_api.get(
+        f"/buckets/{default_bucket.bucket_name}",
+        params={"create_bucket_if_not_exist": False},
+        headers={"Authorization": "Bearer token"},
+    )
     assert response.status_code == status.HTTP_400_BAD_REQUEST, response.text
     assert response.json()["error_code"] == SfsErrorCodes.SFS_BUCKET_NOT_FOUND
     assert response.json()["error_message"] == f"Bucket '{default_bucket.bucket_name}' not found."
 
+    mock_check_access_allow.assert_called_once()
+
 
 @pytest.mark.asyncio
-async def test_delete_bucket(http_client_api, mock_boto_client, default_bucket):
+async def test_delete_bucket(http_client_api, mock_boto_client, default_bucket, mock_check_access_allow):
     mock_boto_client.delete_bucket.return_value = mock.Mock()
 
-    response = await http_client_api.delete(f"/buckets/{default_bucket.bucket_name}")
+    response = await http_client_api.delete(f"/buckets/{default_bucket.bucket_name}", headers={"Authorization": "Bearer token"})
 
     assert response.status_code == status.HTTP_200_OK, response.text
 
     mock_boto_client.delete_bucket.assert_called_once()
     mock_boto_client.delete_bucket.assert_called_once_with(Bucket=default_bucket.bucket_name)
+
+    mock_check_access_allow.assert_called_once()
